Integrate skill chart with Slack command
A skills histogram is now generated and slapped into a Slack formatted
message.  The actual Slack integration has *not* been tested yet.  A
user should be able to enter `/whobot statistics` and be hit with a
beautiful ASCII chart.  The message likely needs to be reformatted, but
should be tested before proceeding futher.